+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Unreleased
 ### v0.0.1-alpha.x
 - Auth filter that match between requested permission and claimed permission
+- Refresh token